pushWriter: correctly propagate errors
In the refactor from 926b9c72f61b5be6bf8d952512f1d0932fbaf898, the error
handling was substantially reworked, and changed the types of errors
returned.

Notably, in the case of a network error, instead of propogating the
error through to return from pushWriter.Write (as previously), it would
be propagated through to pushWriter.Commit - however, this is too late,
since we've already closed the io.Pipe by the time we would have reached
this function. Therefore, we get the generic error message  "io:
read/write on closed pipe" for *every network error*.

This patch corrects this behavior to ensure that the correct error
object is always returned as early as possible:

1. Track the corresponding io.PipeReader for the pushWriter - on any
   error, we CloseWithError the Reader, to ensure that the next Write to
   the Writer returns the desired error. This applies for both network
   errors from request.doWithRetries, as well as for the ErrResets.
2. Remove the error channel - we don't need this any more, errors are
   instead clearly tracked through the CloseWithError function.
3. Always reset the content, even if the offset is already 0. The
   previous patch could fallthrough out of the switch statement, and
   attempt to access a nil response object.